requirements.txt
boto3==1.40.27
botocore==1.40.27
certifi==2025.8.3
charset-normalizer==3.4.3
Events==0.5
idna==3.10
jmespath==1.0.1
opensearch-py==3.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
requests==2.32.5
s3transfer==0.14.0
six==1.17.0
urllib3==2.5.0
